Move playlistnavigation to top and make it nicer (accordion), remove scrollable sides on playlist page, other small adjustments.
