--- src/components/Playlist/Playlist.scss
@@ -1,8 +1,0 @@
-@import "~bootstrapStyles/variables";
-
-.playlistItem {
-  &:hover {
-    color: $brand-primary;
-    cursor: pointer;
-  }
-}
