+++ src/components/PlaylistPage/LessonItem.scss
