+++ src/components/PlaylistPage/PlaylistNavigation.scss
@@ -1,0 +1,3 @@
+.link {
+  cursor: pointer
+}
