+++ src/pages/PlaylistPage.scss
@@ -1,10 +1,3 @@
-.fixed {
-  position: fixed;
-  top: 400px;
-  bottom: 25px;
-}
-
-.scrollable {
-  overflow-y: auto;
-  height: 100%;
+.filter {
+  margin-top: 63px;
 }
